Update assessmentDocument link  lifetime from "long" to "short"

+++ back/src/domains/convention/use-cases/notifications/NotifyBeneficiaryThatAssessmentIsCreated.ts
@@ -55,7 +55,7 @@ export const makeNotifyBeneficiaryThatAssessmentIsCreated = useCaseBuilder(
             role: "beneficiary",
             targetRoute: frontRoutes.assessmentDocument,
             now: today,
-            lifetime: "long",
+            lifetime: "short",
           }),
         },
       },
